+++ model_grids_add.go
@@ -13,8 +13,8 @@ package openapi
 type GridsAdd struct {
 	Id              string `json:"id,omitempty"`
 	NoResultIfLimit bool   `json:"no_result_if_limit,omitempty"`
-	Offset          int32  `json:"offset,omitempty"`
-	Size            int32  `json:"size,omitempty"`
+	Offset          int64  `json:"offset,omitempty"`
+	Size            int64  `json:"size,omitempty"`
 	State           string `json:"state,omitempty"`
 	Rules           string `json:"rules,omitempty"`
 }
